+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### New features
 
+* The `artifact-version` list is now sorted.
 * Add `namespace-aliases` which provides a mapping of the namespace aliases that are in use in the project.
 * Make `find-symbol` able to handle macros.